reformat classes and seperate clases into their own files
