+++ lib/player.rb
@@ -2,9 +2,9 @@
 class Player
   attr_accessor :name, :token
 
-  def initialize(name, token)
+  def initialize(name, token
   self.name = name
     self.token = token
-    "unending string"
+    "unending string
   end
 end
